requirements.txt
bs4
google-api-core==1.22.0
google-auth==1.20.0
google-cloud==0.34.0
google-cloud-core==1.4.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
json5==0.9.1
jsonschema==3.2.0
lxml==4.5.0
numpy==1.18.1
pandas==1.0.1
path==13.1.0
pathlib2==2.3.5
requests==2.22.0
urllib3==1.25.8
nltk==3.4.5
scikit-learn
